+++ doc/CHANGES.rst
@@ -5,6 +5,17 @@ This file contains change information for the current Zope release.
 Change information for previous versions of Zope can be found at
 http://docs.zope.org/zope2/releases/.
 
+2.12.6 (unreleased)
+-------------------
+
+Features Added
+++++++++++++++
+
+
+Bugs Fixed
+++++++++++
+
+
 2.12.5 (2010-04-24)
 -------------------